+++ .flake8
@@ -3,6 +3,6 @@ max-line-length = 88
 extend-ignore=B008
 exclude = .git,__pycache__,old,build,dist
 per-file-ignores =
-    src/db/db_filling_example.py:E501
-    src/vocabulary_builder.py:B008
-    src/db/migration/env.py:F401
+    vocabulary_builder/db/db_filling_example.py:E501
+    vocabulary_builder/main.py:B008
+    vocabulary_builder/db/migration/env.py:F401
