+++ vocabulary_builder/db/db_filling_example.py
@@ -3,7 +3,6 @@ import uuid
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-
 from src.db.models import WordModel
