+++ vocabulary_builder/db/migration/env.py
@@ -4,7 +4,6 @@ from logging.config import fileConfig
 from alembic import context
 from dotenv import load_dotenv
 from sqlalchemy import engine_from_config, pool
-
 from src.db.database import BaseModel
 from src.db.models import ExampleModel, MeaningModel, UserModel, WordModel
