+++ vocabulary_builder/exceptions.py
