+++ vocabulary_builder/static/css/footer.css
