+++ vocabulary_builder/static/css/style.css
