+++ vocabulary_builder/static/css/word_card.css
