+++ vocabulary_builder/static/js/index.js
