+++ vocabulary_builder/static/js/word_card.js
