+++ vocabulary_builder/templates/index.html
@@ -154,7 +154,7 @@
         </div>
     </div>
 
-    <script src="{{ url_for('static', path='/js/index.js') }}"></script>
-    <script src="{{ url_for('static', path='/js/word_card.js') }}"></script>
+    <script vocabulary_builder="{{ url_for('static', path='/js/index.js') }}"></script>
+    <script vocabulary_builder="{{ url_for('static', path='/js/word_card.js') }}"></script>
 </body>
 </html>
